requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
cryptography==41.0.7
python-dotenv==1.0.0
sqlalchemy==2.0.23
pydantic-settings==2.1.0
pydantic==2.5.0
openai==1.3.0
httpx==0.25.2
requests==2.31.0